fix : 전체 폴더 라우팅 수정 & 헤더 수정
