+++ src/app/(letter)/letter/layout.tsx
@@ -1,3 +1,4 @@
+import BottomNavbar from '@/components/common/bottom-navbar';
 import { Header, HeaderCenter } from '@/components/common/header';
 
 export default function PortfolioLayout({
@@ -6,11 +7,12 @@ export default function PortfolioLayout({
   children: React.ReactNode;
 }) {
   return (
-    <>
+    <div className="pb-[66px] pt-[58px]">
       <Header className="border-gray-80 text-lg shadow-none">
         <HeaderCenter className="font-semibold text-gray-10">쪽지</HeaderCenter>
       </Header>
-      <div className="pb-[64px]">{children}</div>
-    </>
+      {children}
+      <BottomNavbar />
+    </div>
   );
 }
