+++ src/app/(my)/my-page/book-mark/page.tsx
@@ -1,0 +1,7 @@
+interface IBookMarkPageProps {}
+
+const BookMarkPage: React.FunctionComponent<IBookMarkPageProps> = () => {
+  return <main></main>;
+};
+
+export default BookMarkPage;
