+++ src/app/(my)/my-page/my-studio/layout.tsx
@@ -1,4 +1,5 @@
 import BackButton from '@/components/common/back-button';
+import BottomNavbar from '@/components/common/bottom-navbar';
 import {
   Header,
   HeaderCenter,
@@ -7,28 +8,26 @@ import {
 } from '@/components/common/header';
 import Icon from '@/components/common/icon';
 
-import React from 'react';
-
-export default function PortfolioLayout({
+export default function MyStudioLayout({
   children,
-}: {
+}: Readonly<{
   children: React.ReactNode;
-}) {
-  // id로 닉네임 받아오기
+}>) {
   return (
     <div className="pb-[66px] pt-[58px]">
-      <Header className="border-none shadow-none">
+      <Header>
         <HeaderLeft>
           <BackButton>
             <Icon id="back-icon" className="size-[32px] text-gray-40" />
           </BackButton>
         </HeaderLeft>
-        <HeaderCenter>닉네임원 닉네임원</HeaderCenter>
+        <HeaderCenter>마이 스튜디오</HeaderCenter>
         <HeaderRight>
-          <div className="size-[32px]"></div>
+          <Icon id="notification-icon" className="h-6 w-6 text-gray-40" />
         </HeaderRight>
       </Header>
-      <main className="relative px-[16px] pb-[79px]">{children}</main>
+      {children}
+      <BottomNavbar />
     </div>
   );
 }
